mctp-astpcie-discovery: Add simple app to run on HW
Add a simple application that uses libmctp API to communicate with
ast2600-mctp driver.

The provided program allows to verify both TX and RX flow based on MCTP
Endpoint Discovery.

It handles and sends responses to the basic control commands like:
- Prepare for Endpoint Discovery,
- Endpoint Discovery,
- Get Endpoint Id,
- Set Endpoint Id.

It also generates and receives a response for Discovery Notify command.

Tested:
On Mehlow server platform with AST2600 EVB as PCIe endpoint and SPS FW
as MCTP Bus Owner.

Change-Id: I5b004854b677c1714d819918741490c68ca62899
Signed-off-by: Iwona Winiarska <iwona.winiarska@intel.com>