libmctp: handle memory allocation for message binding private data
This commit introduces memory allocation for binding private data.
Earlier mechanism was passing a void* pointer to the library;
however if the caller de-allocated memory for the binding private
data, it could result in segfaults. Thus this commit introduces
memory for the binding private data in the library and we keep
the copy of the value passed to us.

Also added test cases to test message_binding_private.

Change-Id: I88ee1db2a1153a495415dbe8968becb159bde54a
Signed-off-by: Sumanth Bhat <sumanth.bhat@intel.com>