astpcie: Modify tx/rx methods
The proper PCIe header is a crucial thing in case of MCTP over PCIe
communication. Even a single invalid frame may cause hardware to stop
working correctly.

In this change we're simplifying the existing implementation, adding
various helpers to get/set the relevant header fields.
This allows us to hide the "implementation details" that deal with
endianess conversion from the callers.

Change-Id: I3666ea21da5dfab3b56730266b841e7cfa5fe351
Signed-off-by: Iwona Winiarska <iwona.winiarska@intel.com>