utils: Add mctp-astpcie-register-type app
This adds simple application that tests API methods that register and
unregister type handlers. The test registers for MCTP Control messages
and waits for one. After receiving the message successfully it registers
for VDPCI with some artificial vendor ID and vendor type. Type handler is
unregistered after receiving a suitable message.

Tested:
A platform with AST2600 EVB and MCTP Emulator.

Change-Id: I291756003f543182eccb53a81e4167e3bae272db
Signed-off-by: Oleszkiewicz, Martyna <martyna.oleszkiewicz@intel.com>