astpcie: Fix length field value encoding
libmctp uses the length[9:0] value from PCIe header as-is, without
checking if it is equal 0 which corresponds with the maximum data
payload size of PCIe frame (1024 DW). In this case, use of zero for
payload size calculation will cause integer overflow.

Let's fix this, by checking length[9:0] value - if it's equal to 0, we will
use the maximum data payload size for the further calculations.

There is nothing to change for TX path - since we already used macros
that will set the maximum data payload size correctly.

Let's add an assert to prevent setting incorrect packet size value.

Tested on HW. No regression was found.

Reported-by: Hareesh Khattri <hareesh.khattri@intel.com>
Signed-off-by: Iwona Winiarska <iwona.winiarska@intel.com>
Change-Id: I8373a32ed786d5f7b6f375de2faabf55f6997106