Fix compiler warnings when I2C_M_HOLD is undefined
If the flag I2C_M_HOLD is undefined, parts of the code
need to be conditionally excluded. However, this does not
happen in a few cases.

This change wraps portions of the code that compile
only if the flag is defined.

Tested: No functional changes. Code compiles in both cases:
  - Macro I2C_M_HOLD is defined.
  - Macro I2C_M_HOLD is undefined.

Change-Id: Iad7c988b1835f2c9726f80aff41bfcc9a33d5a9c
Signed-off-by: Sandino Flores <sandino.flores.moreno@intel.com>